A* algorithm
The A* algorithm finds the shortest path between 2 points in a grid. accordingly to a formula f(n) = h(n) + g(n), where h is a guess of a cost of going from a node to the final node. For h(n) we used the Manhattan distance. g(n) is the cost to get to node n. In the end the shortest path is the collection of nodes with the lowest f(n).

This program runs using pygame. 

Shoutout goes to Tech with Tim on youtube.