Database
In this project we create a database and database GUI with python using its tkinter library. What can you do with this program? You can create a database, add entries to the database, delete entries from the database. A feature to update an entry in the database is being looked into. It exists but not function yet. 
Updates coming soon